crawler/requirements.txt
autoit==0.1
Automat==0.7.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
colorlog==3.1.4
colormap==1.0.1
constantly==15.1.0
cryptography==2.3
cssselect==1.0.3
Cython==0.29.7
decorator==4.3.0
defusedxml==0.6.0
easydev==0.9.37
flex==6.13.2
future==0.16.0
genson==1.0.1
get-mac==0.8.2
html-testRunner==1.1.2
http-basic-auth==1.2.0
hyperlink==18.0.0
idna==2.7
image==1.5.27
incremental==17.5.0
jsonpath-ng==1.4.3
jsonpointer==1.14
jsonschema==3.0.0a3
libpcap==1.10.0b10
lxml==4.2.4
MarkupSafe==1.0
netaddr==0.7.19
netifaces==0.10.9
numpy==1.15.0
oauthlib==3.1.0
pandas==0.23.4
parsel==1.5.0
pexpect==4.6.0
Pillow==6.0.0
ply==3.11
prompt-toolkit==1.0.15
ptyprocess==0.6.0
py-webdriver-manager==0.0.1
pyaml==19.12.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyAutoIt==0.6.0
pycparser==2.18
pydevd==1.4.0
PyDispatcher==2.0.5
Pygments==2.2.0
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pyrsistent==0.14.9
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
pywin32==227
PyYAML==3.13
queuelib==1.5.0
randmac==0.0.2
reports==0.3.1
requests==2.20.0
requests-file==1.4.3
requests-oauthlib==1.3.0
RESTinstance==1.0.0rc4
rfc3987==1.3.8
robotframework==3.0.4
robotframework-debuglibrary==1.1.4
robotframework-pabot==0.44
robotframework-requests==0.5.0
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==3.1.1
robotremoteserver==1.1
scapy==2.4.3
selenium==3.14.0
selenium-requests==1.3
service-identity==17.0.0
six==1.11.0
strict-rfc3339==0.7
tldextract==2.2.0
tqdm==4.23.1
Twisted==19.7.0
tzlocal==1.5.1
urllib3==1.23
validate-email==1.3
virtualenv==16.7.2
w3lib==1.19.0
waiting==1.4.1
wcwidth==0.1.7
webcolors==1.8.1
webium==1.2.1
Whoosh==2.7.4
winpcapy==1.0.2
zope.interface==4.5.0
